+++ projects/fhi-angular-highcharts/CHANGELOG.md
@@ -1,11 +1,12 @@
 # Unreleased
 
-> Feb 14, 2025
+> Mar 03, 2025
 
 * :tada: **Enhancement** Update to new control flow syntax in templates [(#782)](https://github.com/folkehelseinstituttet/Fhi.Frontend.Demo/pull/782)
-* :tada: **Enhancement** Update to Angular v19 [(#800)](https://github.com/folkehelseinstituttet/Fhi.Frontend.Demo/pull/800)
+* :boom: **Breaking** Update to Angular v19 [(#800)](https://github.com/folkehelseinstituttet/Fhi.Frontend.Demo/pull/800)
 * :bug: **Bugfix** Make SVG-download work also for maps [(#803)](https://github.com/folkehelseinstituttet/Fhi.Frontend.Demo/pull/803)
 * :bug: **Bugfix** Make sure navigation group id can't become undefined [(#804)](https://github.com/folkehelseinstituttet/Fhi.Frontend.Demo/pull/804)
+* :boom: **Breaking** Improve decimal handling. NB! Remove support for `unit.decimals = null`, since the type has always just been `number`  [(#814)](https://github.com/folkehelseinstituttet/Fhi.Frontend.Demo/pull/814)
 
 ## 5.3.0